Use libbitcoin compatible libsecp256k1
To get AES256 decryption, we'll need routines in libbitcoin and the known version we can compile has it's own libsecp256k1 library built in. This change modifies the use of the secp256k1 library to its prior API that is compatible with the libbitcoin version we use.

Note: This change does not include necessary updated binary files.