Add new repoFileGetText and repoFileGetData routines
These load a repo off disk and decrypt it in native code. Saves a roundtrip through the bridge. This also implements the logic for a RepoUnionFolder allowing optional arguments for fallback and whiteout paths so failure to load master will load fallback if whiteout doesn't exist.

Bridge only implemented for iOS, not Android yet

Implement all routines on a separate thread to not block the UI main thread.
Add commented out lines to Makefiles and recipes to build debug versions of C++ files.
Note: binaries are not updated in repo as of this commit. Final commit with release binaries needs to be made.