Compile secp256k1 as an XCFramework
This will allow everything to work normally on an M1 Mac.

To make this happen, we have to pull all other sources out of the native build, leaving just secp256k1 by itself. We put the sources in the Android folder because Android cares about file locations, but iOS does not.

Once that is done, we can adjust the iOS build instructions to be SDK-aware, and to bundle the output as an XCFramework.